+++ src/Factory/InformationFactory.php
@@ -15,4 +15,9 @@ interface InformationFactory
     public function createBlocks(string $local_path_to_code): array;
 
     public function getRepositoryRequirements(): RepositoryRequirements;
+
+    /**
+     * @return array An array of Fully Qualified Classnames of the Information block types that the factory can produce
+     */
+    public function listAvailableInformationBlocks(): array;
 }