+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 
 ## Version 0.0.0
 ### Changes
+* Added new interface Repository and InformationBlock
 * Adjust namespace
 * Added Exception
 * Added initial Models and Interfaces